scanning a promo QR code works now, shows the details of an event to the user.
HOW SCANNING WORKS:
a promo QR code contains the EVENT ID of an event with the PREFIX promo, e.g. promo12345abcdefg.
If the scan detects this prefix it triggers promo behaviour and goes to the event details if applicable.

a check-in QR code contains just the event id with no prefix. if there is no prefix then the user is sent to a check-in page and is added to a list of checked in attendees on the event.

i did not have enough time to actually add room in the app for both the promo and check-in qr codes.

i've added some test promo/check-in qr codes to the files, under the folder testQRCodes.