Created Tests for Admin Dashboard
On my pc, it fails because the events and profiles arent loading. When I was first working on it, they were loading. Then they stopped. I have no idea why they are not appearing. They are appearing in the normal app just fine. I have to move onto other things. This will have to do for now.