Notification Switch on EditProfileActivity is functional
When toggled, if the user does not have notifactions enabled it will open an activity for them to enable it through the android settings. As of right now, we can't toggle the switch, because the "App Does not send any notifications". Once the app has notifications built in this switch will be toggleable.