custom promotional qr code now functions properly
if you upload a custom qr code,
it now displays in the qr management section of the app.

also tweaked geolocation stuff so that it first checks server side if the user wants to be geolocated, then asks for permissions (so if they have geolocation off the app makes no attempt to grab location, even if they may have given the app permission to do so) this is probably the best way to do it.